fix(projects): use cloudinary to host project images
Previously project images come from the notion and now come from the
cloudinary. Cloudinary is a awesome image and video host platform.

Eliminate the `ProjectImage` type, as images are provided in the format of
a URL rather than as a file type from Notion.

Configure cloudinary protocol and hostname on the next.config.mjs file to
allow cloudinary in this project.