chore!: transform the repository to monorepo with `turborepo`
BREAKING CHANGE: This repository is transformed to the monorepo with
`turborepo` and `pnpm`.

Folders and files structure are changed according to a monorepo. Now
it's a monorepo with two workspace folders: apps and packages. These
workspaces are manage by the `pnpm`. And `turborepo` helps to work
faster in a monorepo.